% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{add_lr_to_history}
\alias{add_lr_to_history}
\title{Modifies the keras history object during training to
include the learning rates of each epoch as a metric}
\usage{
add_lr_to_history(verbose = TRUE)
}
\description{
Modifies the keras history object during training to
include the learning rates of each epoch as a metric
}
