% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{as.fraction}
\alias{as.fraction}
\title{Farey sequence approximation algorithm to fractions, WIP}
\usage{
as.fraction(x, n = 1000, delta = 0.01, return_precision = FALSE)
}
\description{
Farey sequence approximation algorithm to fractions, WIP
}
