% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_and_compile}
\alias{build_and_compile}
\title{Convencience to build a keras model with given specification.}
\usage{
build_and_compile(
  input,
  output,
  optimizer = "adam",
  loss = "mse",
  metric = "acc"
)
}
\arguments{
\item{input}{tensor (or keras layer_input) defining model input.}

\item{output}{tensor (or keras layer_input) defining model output.}

\item{optimizer}{keras optimizer (character or callable) to use.}

\item{loss}{loss function (character or callable).}

\item{metric}{metric to score model with.}
}
\description{
Convencience to build a keras model with given specification.
}
