% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{calculate_divisor_pairs}
\alias{calculate_divisor_pairs}
\title{Calculate all divisor pairs for a given number}
\usage{
calculate_divisor_pairs(x)
}
\description{
Calculate all divisor pairs for a given number
}
