% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{callback_save_optimizer_weights}
\alias{callback_save_optimizer_weights}
\title{wrapper around R6 method callback for saving epoch optimizer weights}
\usage{
callback_save_optimizer_weights()
}
\description{
wrapper around R6 method callback for saving epoch optimizer weights
}
