% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-test.R
\name{ex_model}
\alias{ex_model}
\title{Dummy 3 layer dense model to quickly test on mnist (or other) data}
\usage{
ex_model(input_shape = list(784), compile = TRUE)
}
\description{
Dummy 3 layer dense model to quickly test on mnist (or other) data
}
