% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1_score_binary}
\alias{f1_score_binary}
\title{Computes F1 score for a binary classifier given the logit and label vectors.}
\usage{
f1_score_binary(logits, labels)
}
\description{
Computes F1 score for a binary classifier given the logit and label vectors.
}
