% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_list}
\alias{is_list}
\title{Is x explicitly an object of type list}
\usage{
is_list(x)
}
\arguments{
\item{x}{input object}
}
\description{
Is x explicitly an object of type list
}
\examples{
is_list(c(1,2,3))
is_list(list(99))
}
