% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_argmax}
\alias{layer_argmax}
\title{A differentiable version of tf$one_hot(tf$argmax(x, -1L), nchannels)}
\usage{
layer_argmax(object, ...)
}
\description{
A differentiable version of tf$one_hot(tf$argmax(x, -1L), nchannels)
}
