% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glu.R
\name{layer_glu_block}
\alias{layer_glu_block}
\title{Returns a number of connected Gated Linear Units. See `layer_glu()`}
\usage{
layer_glu_block(
  object,
  filters,
  num_layers = 3,
  kernel_size = 3,
  kernel_initializer = "glorot_normal",
  kernel_regularizer = NULL,
  bias_initializer = "zeros",
  bias_regularizer = NULL,
  name = NULL,
  trainable = TRUE
)
}
\description{
Returns a number of connected Gated Linear Units. See `layer_glu()`
}
