% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-scale.R
\name{mean_normalize}
\alias{mean_normalize}
\title{Creates a distribution that will have values between `(-1, 1)` with `μ = 0`.}
\usage{
mean_normalize(x)
}
\description{
Creates a distribution that will have values between `(-1, 1)` with `μ = 0`.
}
