% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{mnist_dataset}
\alias{mnist_dataset}
\title{Prepare an MNIST tfdatasets object and returns `c(ds, val_ds)`,
where `ds` and `val_ds` are tf.data.Dataset objects.  Each
dataset object has a tuple of `(x, y)` pairs of shape `(examples, 784)`.}
\usage{
mnist_dataset()
}
\description{
Prepare an MNIST tfdatasets object and returns `c(ds, val_ds)`,
where `ds` and `val_ds` are tf.data.Dataset objects.  Each
dataset object has a tuple of `(x, y)` pairs of shape `(examples, 784)`.
}
