% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{padded_imdb_data}
\alias{padded_imdb_data}
\title{Convenience wrapper around `keras::dataset_imdb()` which pads
input seqequences for use with a keras model in addition to
preprocessing already applied in keras.}
\usage{
padded_imdb_data(vocab_size = 10000, max_length = 150, trunc_type = "post")
}
\description{
Convenience wrapper around `keras::dataset_imdb()` which pads
input seqequences for use with a keras model in addition to
preprocessing already applied in keras.
}
