% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{pow2_range}
\alias{pow2_range}
\title{Return powers of 2 within a specified range}
\usage{
pow2_range(min, max)
}
\arguments{
\item{min}{minimum of desired range}

\item{max}{maximum of desired range}
}
\description{
Return powers of 2 within a specified range
}
