% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{primes_n}
\alias{primes_n}
\title{Generate prime numbers to a specified limit.}
\usage{
primes_n(limit)
}
\arguments{
\item{limit}{Limit to generate up to.}
}
\description{
Not particularly efficient, but gets the job done.
Try to keep it under 100000
}
