% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.R
\name{resblock_batchnorm_bottle_1d_v2}
\alias{resblock_batchnorm_bottle_1d_v2}
\title{resblock with bottleneck from the Exploring Normalization FB paper (01/2017)
modified as in: https://www.kaggle.com/h030162/version1-0-9696}
\usage{
resblock_batchnorm_bottle_1d_v2(
  x,
  filters = 64L,
  kernel_size = 3L,
  downsample = TRUE,
  renorm = FALSE
)
}
\description{
resblock with bottleneck from the Exploring Normalization FB paper (01/2017)
modified as in: https://www.kaggle.com/h030162/version1-0-9696
}
