% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{return_decimal}
\alias{return_decimal}
\title{Remove integer portion of a numeric and return only decimals}
\usage{
return_decimal(x)
}
\description{
Remove integer portion of a numeric and return only decimals
}
