% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{return_decimal2}
\alias{return_decimal2}
\title{Remove integer portion of numeric and return decimals.
Slightly faster than return_decimal.}
\usage{
return_decimal2(x)
}
\description{
Remove integer portion of numeric and return decimals.
Slightly faster than return_decimal.
}
