% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curves.R
\name{sample_lr_values_log_scale}
\alias{sample_lr_values_log_scale}
\title{Randomly sample values for learning rate parameter `alpha` on a log scale}
\usage{
sample_lr_values_log_scale(n_grid_points = 10, lo = 1e-07, hi = 1)
}
\arguments{
\item{n_grid_points}{integer, number of values to sample}

\item{lo}{minimum learning rate value to return}

\item{hi}{maximum learning rate value to return}
}
\value{
vector of `alpha` samples within range [log(lo, 10), log(hi, 10)]
Potentially run this after the lr_range_test to narrow?
}
\description{
Computes minimum values on a log scale as `a <- log(lo, 10)`, and
 `b <- log(hi, 10)`.  Calls to `runif()` are scaled to the appropriate
 range by `a` and are checked to be out of bounds against `-b`.  Values
 of `alpha` are computed by raising `10` to the sampled value power.
}
