% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-scale.R
\name{scaler}
\alias{scaler}
\title{Perform feature scaling on a vector, matrix, data.frame, or tibble}
\usage{
scaler(x, type = "standard", margin = 2L, ...)
}
\arguments{
\item{x}{One of (data.frame, matrix, tibble)}

\item{type}{string indicating the type of supported scaling function.  Supported
types: c(standardization, mean scaling, minmax scaling, unit vector scaling, sqrt)
default: standardize}

\item{margin}{margin to apply over (default: 2L)}
}
\description{
Perform feature scaling on a vector, matrix, data.frame, or tibble
}
