% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\name{split_numeric}
\alias{split_numeric}
\title{Split integer from decimal part of a numeric primitive.
Note: only the integer portion retains original sign(x)}
\usage{
split_numeric(x)
}
\description{
Split integer from decimal part of a numeric primitive.
Note: only the integer portion retains original sign(x)
}
