% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-scale.R
\name{sqrt_scale}
\alias{sqrt_scale}
\title{Scale values based on `sqrt(sum(x^2) / max(1, length(x) - 1L))`}
\usage{
sqrt_scale(v)
}
\description{
Scale values based on `sqrt(sum(x^2) / max(1, length(x) - 1L))`
}
