% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-test.R
\name{test_mnist_mlp}
\alias{test_mnist_mlp}
\title{Trains a simple deep NN on the MNIST dataset.}
\usage{
test_mnist_mlp()
}
\description{
Gets to 98.40% test accuracy after 20 epochs (there is *a lot* of margin for
parameter tuning).
}
