% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning-curves.R
\name{train_set_size_run}
\alias{train_set_size_run}
\title{Complete a set of training runs given an integer set of training set sizes.}
\usage{
train_set_size_run(
  model_fn,
  x_train,
  y_train,
  type = "linear",
  epochs = 25,
  n_elems = dim(x_train)[1],
  linear_factor = 10,
  train_sizes = NULL,
  min_train_size = NULL,
  n_runs = NULL
)
}
\description{
Plots Jtrain and Jcv loss and accuracy curves
}
